listener_messages/requirements.txt
kafka-python==2.0.2
PyJWT==2.3.0
requests==2.27.1
